Restructure config package
